+++ lib/cdt/dtopen.c
@@ -23,7 +23,6 @@ Dt_t* dtopen(Dtdisc_t* disc, Dtmethod_t* meth)
 	dt->meth = NULL;
 	dt->disc = NULL;
 	dtdisc(dt, disc);
-	dt->type = DT_MALLOC;
 	dt->nview = 0;
 	dt->view = dt->walk = NULL;
 	dt->user = NULL;