+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Criterion unit tests have been removed and migrated to Pytest. This is
   primarily relevant to downstream packagers of Graphviz. #2443
 - **Breaking**: `Dtdisc_t.memoryf` and its associated macros has been removed.
+- **Breaking**: The `Dt_t.type` field has been removed.
+- **Breaking**: The `dtfound`, `DT_FOUND`, `dtleast`, and `dtmost` macros have
+  been removed.
+- The nrtmain.c test program has been removed from the portable tarball.
 
 ### Fixed